[Divide by 0 Error] norm check migrate to c++
